chore: extract isSameOrigin into shared helper and apply to all destructive endpoints (#87)
destructive POST/DELETE 全般に Origin check を defense-in-depth として展開。

- src/lib/auth/csrf.ts: PR-C #82 の isSameOrigin を共通 helper に抽出
- /api/scrim/finalize / /api/scrim/[id] / /api/auth/dev/test-login にも適用
- 既存 /api/account/delete は新 helper を import するように
